Fixed WBP_Stats Bug & Begun Stat Drop Development
WBP_Stats Bug is when Stat Points would be consumed when the Player allocates stat points and closes the Stats Menu afterwards without confirmation. This causes Stat Points to be consumed without upgrading Stats. Fixed this by only allowing the Stats Menu to be opened after every wave if the Player has Stat Points. Additionally, Player cannot close the Stats Menu until the confirmation button is pressed.
Stat Drop currently is a pickup-able item that can be interacted with the Player Character and increment the Player's MODStatPoints. It does not have a visual representation yet.